[Refactor] @given-dragon 리뷰 반영
