+++ src/main/kotlin/com/whatever/domain/user/dto/PostUserProfileRequest.kt
@@ -2,6 +2,7 @@ package com.whatever.domain.user.dto
 
 import io.swagger.v3.oas.annotations.media.Schema
 import jakarta.validation.constraints.NotBlank
+import jakarta.validation.constraints.Pattern
 import jakarta.validation.constraints.Size
 import java.time.LocalDate
 
@@ -9,6 +10,7 @@ data class PostUserProfileRequest(
     @Schema(description = "닉네임")
     @field:NotBlank(message = "닉네임은 공백일 수 없습니다.")
     @field:Size(min = 3, max = 10, message = "닉네임은 3~10자 이내로 입력해주세요.")
+    @field:Pattern(regexp = "^[가-힣a-zA-Z0-9]+$")
     val nickname: String,
     @Schema(description = "YYYY-MM-DD 형식의 생일")
     val birthday: LocalDate,
